feat:rename mapreduce
