+++ app/index_platform/analyzer/init.go
