+++ app/index_platform/analyzer/token_test.go
