+++ app/index_platform/client/map.go
