+++ app/index_platform/client/reduce.go
