+++ app/index_platform/service/mapreduce_test.go
